% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Bundle a Shiny App with WebR Bundle}
\usage{
build(appdir = getwd(), outdir = file.path(getwd(), "dist"), parallel = 4)
}
\arguments{
\item{appdir}{The directory of the Shiny App}

\item{outdir}{The directory to output the bundled app (default: dist)}

\item{parallel}{The number of packages to bundle in parallel (default: 4)}
}
\description{
Bundle a Shiny App with WebR Bundle
}
