% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_from_source}
\alias{install_from_source}
\title{Install webr-bundle from source}
\usage{
install_from_source()
}
\description{
This function installs webr-bundle from source.
It is only supported on macOS and Linux and
requires Rust to be installed. If Rust is not
installed, it will be installed automatically.
}
