% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve}
\alias{serve}
\title{Serve a Shiny App with WebR Bundle}
\usage{
serve(outdir = file.path(getwd(), "dist"), port = 8080)
}
\arguments{
\item{outdir}{The directory where the bundle is located (default: dist)}

\item{port}{The port to serve the app on (default: 8080)}
}
\description{
Serve a Shiny App with WebR Bundle
}
